# Add dot product benchmarks, improve performance