@danieldk 2021-05-06
Clippy emits a lot of

```
unsafe function's docs miss `# Safety` section
```

Now that the raw vectorized dot products are public (for benchmarking). I think many of these functions are provably safe. I'll update the PR tonight.