@sebpuetz 2021-05-06
I'll take a look later today, too! I gave it a quick glance earlier and on my machine there's also nice improvements :)

There was some superfluous `unsafe` block involved in one of the benches too, think it was the unvectorized one!